pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sciagent"
version = "2.0.0"
description = "Advanced Multi-Agent Scientific & Software Development System"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "SciAgent Team"}
]
keywords = ["ai", "science", "research", "multi-agent", "experimentation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.0",
    "openai>=1.50.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "asyncio>=3.4.3",
    "aiohttp>=3.9.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "datasets>=2.14.0",
    "transformers>=4.30.0",
    "mcp>=0.9.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
sciagent = "sciagent.cli:cli"

[project.urls]
Homepage = "https://github.com/yourusername/sciagent"
Documentation = "https://sciagent.readthedocs.io"
Repository = "https://github.com/yourusername/sciagent"
Issues = "https://github.com/yourusername/sciagent/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
